// hrms-frontend/src/app/modules/employee/hosonhansu/model/hopdonglaodongUI.ts
export interface HopdonglaodongUI {
    hdldId?: number;
	nsID?: number;
	vtriId?: number;
	donviId?: number;
	nsQdndung?: any;
	nsQdnoidungId?: number;
	sohopdong?: String;
	nguoisudung?: String;
	chucvuky?: String;
	tenfile?: String;
	fileExtend?: String;
	//fileAttach?: any;
	isFileQd?: boolean;
	ngaykyhopdong?: Date;
	noidunghdong?: String;
	loaiHD_ID?: number;
	loaiHD?: String;
	ngaychinhthuc?: Date;
	ngayhethan?: Date;
	// private Long vtri_ID;
	// private String vtricdanh;
	congviec?: String;
	// private Double luongtra;
	noidungkhac?: String;
	isHientai?: boolean;
	koDongBHXH?: boolean;
	isChangeFileAttach?: boolean;
}